client: add ReservationTr
