+++ client/src/pages/index.tsx
@@ -1,22 +1,12 @@
 import RestaurantCard from "@/components/RestaurantCard";
 import { contractAddresses } from "@/config/constants";
 import { s2RnftAbi, stake2ReserveAbi } from "@/generated";
-import { Box, HStack, Text, Image, Button, VStack, Table, TableCaption, TableContainer, Tbody, Td, Tr, Th } from "@chakra-ui/react";
+import { HStack, Text, VStack, Table, TableContainer, Tbody, Tr, Th } from "@chakra-ui/react";
 import Head from "next/head";
 import { useAccount, useConnect, useDisconnect, useReadContract } from "wagmi";
-import NextLink from "next/link";
 import S2RLayout from "@/components/S2RLayout";
 import { zeroAddress } from "viem";
-
-function getDateTimeFromUnixTimestamp(timestamp: number): string {
-    const date = new Date(timestamp * 1000);
-    const month = (date.getUTCMonth() + 1).toString().padStart(2, "0");
-    const day = date.getUTCDate().toString().padStart(2, "0");
-    const hours = date.getUTCHours().toString().padStart(2, "0");
-    const minutes = date.getUTCMinutes().toString().padStart(2, "0");
-
-    return `${month}-${day} ${hours}:${minutes}`;
-}
+import ReservationTr from "@/components/ReservationTr";
 
 export default function Home() {
     const { connectors, connect } = useConnect();
@@ -37,13 +27,6 @@ export default function Home() {
         args: [address ? address : zeroAddress],
     });
     console.log("nftIds: ", nftIds);
-    const { data: reservationData } = useReadContract({
-        address: contractAddresses.S2R,
-        abi: stake2ReserveAbi,
-        functionName: "getReservationData",
-        args: [nftIds ? nftIds[0] : BigInt(0)],
-    });
-    console.log("reservationData: ", reservationData);
 
     return (
         <>
@@ -65,7 +48,7 @@ export default function Home() {
                         Reservations
                     </Text>
                     {isConnected &&
-                        (reservationData ? (
+                        (nftIds ? (
                             <TableContainer w="80%" mt={-2}>
                                 <Table size="sm">
                                     <Tbody>
@@ -80,17 +63,9 @@ export default function Home() {
                                                 Guests
                                             </Th>
                                         </Tr>
-                                        <Tr>
-                                            <Td fontSize={16} borderColor="gray.900">
-                                                {reservationData.shopName}
-                                            </Td>
-                                            <Td fontSize={16} borderColor="gray.900">
-                                                {getDateTimeFromUnixTimestamp(Number(reservationData.startingTime))}
-                                            </Td>
-                                            <Td fontSize={16} borderColor="gray.900">
-                                                {reservationData.guestCount.toString()}
-                                            </Td>
-                                        </Tr>
+                                        {nftIds.map((nftId, i) => (
+                                            <ReservationTr nftId={nftId} />
+                                        ))}
                                     </Tbody>
                                 </Table>
                             </TableContainer>
